pyproject.toml
[project]
name = "climo-data-importer"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.10,<3.14"
dependencies = [
    "pycds @ git+https://github.com/pacificclimate/pycds.git@i-229-multi-climo-normals",
    "testing-postgresql (==1.3.0)",
    "pytest-mock (>=3.15.1,<4.0.0)",
    "pytest-alembic (>=0.12.1,<0.13.0)"
]

[tool.poetry]
packages = [{include = "climo_data_importer", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
